// frontend/packages/common/lib/enums/FileType.ts
export enum FileType {
  ApplicationDoc = 'application/msword',
  ApplicationDocx = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
  ApplicationOds = 'application/vnd.oasis.opendocument.spreadsheet',
  ApplicationOdt = 'application/vnd.oasis.opendocument.text',
  ApplicationPdf = 'application/pdf',
  ApplicationXls = 'application/vnd.ms-excel',
  ApplicationXlsx = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
  ImageDwg = 'image/vnd.dwg',
  ImageDxf = 'image/vnd.dxf',
  ImageGif = 'image/gif',
  ImageJpeg = 'image/jpeg',
  ImagePng = 'image/png',
  ImageSvg = 'image/svg+xml',
  ImageTiff = 'image/tiff',
  MessageEml = 'message/rfc822',
  TextCsv = 'text/csv',
  Video = 'video/*',
}